+++ openeogeotrellis/backend.py
@@ -468,10 +468,20 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                             "description": "Specifies the filename prefix when outputting multiple files. By default, depending on the context, 'OpenEO' or a part of the input filename will be used as prefix.",
                             "default": None,
                         },
-                        "separate_asset_per_band": {  # TODO: should be a boolean that defaults to false
-                            "type": "string",
+                        "separate_asset_per_band": {
+                            "type": "boolean",
                             "description": "Set to true to write one output tiff per band. If there is a time dimension, the files will be split on time as well.",
+                            "default": False,
+                        },
+                        "filepath_per_band": {
+                            "type": "array",
+                            "description": "Specify an array with the same amount of bands, to give each band it's separate path. Subfolders are allowed.",
                             "default": None,
+                        },
+                        "attach_gdalinfo_assets": {
+                            "type": "boolean",
+                            "description": "Attaches *_gdalinfo.json files to the output results.",
+                            "default": False,
                         },
                     },
                 },
@@ -519,6 +529,12 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                     "title": "JavaScript Object Notation (JSON)",
                     "description": "JSON is a generic data serialization format. Being generic, it allows to represent various data types (raster, vector, table, ...). On the other side, there is little standardization for complex data structures.",
                     "gis_data_types": ["raster", "vector"],
+                    "parameters": {},
+                },
+                "GeoJSON": {
+                    "title": "GeoJSON",
+                    "description": "GeoJSON is supported to export vector cube data. GeoJSON is always in EPSG:4326. ",
+                    "gis_data_types": ["vector"],
                     "parameters": {},
                 },
                 "CSV": {
@@ -1882,38 +1898,39 @@ class GpsBatchJobs(backend.BatchJobs):
         if get_backend_config().setup_kerberos_auth:
             setup_kerberos_auth(self._principal, self._key_tab, self._jvm)
 
+        memOverheadBytes = as_bytes(executor_memory_overhead)
+        jvmOverheadBytes = as_bytes("128m")
+
+        # By default, Python uses the space reserved by `spark.executor.memoryOverhead` but no limit is enforced.
+        # When `spark.executor.pyspark.memory` is specified, Python will only use this memory and no more.
+        python_max = job_options.get("python-memory", get_backend_config().default_python_memory)
+        if python_max is not None:
+            python_max = as_bytes(python_max)
+            if "executor-memoryOverhead" not in job_options:
+                memOverheadBytes = jvmOverheadBytes
+                executor_memory_overhead = f"{memOverheadBytes // (1024 ** 2)}m"
+        else:
+            # If python-memory is not set, we convert most of the overhead memory to python memory
+            # this in fact duplicates the overhead memory, we should migrate away from this appraoch
+            python_max = memOverheadBytes - jvmOverheadBytes
+            executor_memory_overhead = f"{memOverheadBytes // (1024 ** 2)}m"
+
+        if as_bytes(executor_memory) + as_bytes(executor_memory_overhead) + python_max > as_bytes(
+                get_backend_config().max_executor_or_driver_memory
+        ):
+            raise OpenEOApiException(
+                message=f"Requested too much executor memory: "
+                        + f"{executor_memory} + {executor_memory_overhead} + {python_max // (1024 ** 2)}m, "
+                        + f"the max for this instance is: {get_backend_config().max_executor_or_driver_memory}",
+                status_code=400,
+            )
+
         if isKube:
             # TODO: get rid of this "isKube" anti-pattern, it makes testing of this whole code path practically impossible
 
             # TODO: eliminate these local imports
             from kubernetes.client.rest import ApiException
 
-
-            memOverheadBytes = as_bytes(executor_memory_overhead)
-            jvmOverheadBytes = as_bytes("128m")
-
-            # By default, Python uses the space reserved by `spark.executor.memoryOverhead` but no limit is enforced.
-            # When `spark.executor.pyspark.memory` is specified, Python will only use this memory and no more.
-            python_max = job_options.get("python-memory", None)
-            if python_max is not None:
-                python_max = as_bytes(python_max)
-                if "executor-memoryOverhead" not in job_options:
-                    memOverheadBytes = jvmOverheadBytes
-                    executor_memory_overhead = f"{memOverheadBytes//(1024**2)}m"
-            else:
-                # If python-memory is not set, we convert most of the overhead memory to python memory
-                python_max = memOverheadBytes - jvmOverheadBytes
-                executor_memory_overhead = f"{memOverheadBytes//(1024**2)}m"
-
-            if as_bytes(executor_memory) + as_bytes(executor_memory_overhead) + python_max > as_bytes(
-                    get_backend_config().max_executor_or_driver_memory
-            ):
-                raise OpenEOApiException(
-                    message=f"Requested too much executor memory: "
-                    + f"{executor_memory} + {executor_memory_overhead} + {python_max//(1024**2)}m, "
-                    + f"the max for this instance is: {get_backend_config().max_executor_or_driver_memory}",
-                    status_code=400,
-                )
 
             api_instance_custom_object = kube_client("CustomObject")
             api_instance_core = kube_client("Core")
@@ -2115,6 +2132,8 @@ class GpsBatchJobs(backend.BatchJobs):
             submit_script = "submit_batch_job_spark3.sh"
             script_location = pkg_resources.resource_filename("openeogeotrellis.deploy", submit_script)
 
+            image_name = job_options.get("image-name", os.environ.get("YARN_CONTAINER_RUNTIME_DOCKER_IMAGE"))
+
             extra_py_files=""
             if len(py_files)>0:
                 extra_py_files = "," + py_files.join(",")
@@ -2201,11 +2220,15 @@ class GpsBatchJobs(backend.BatchJobs):
                 args.append(str(job_work_dir / UDF_PYTHON_DEPENDENCIES_ARCHIVE_NAME))
                 args.append(os.environ.get("OPENEO_PROPAGATABLE_WEB_APP_DRIVER_ENVARS", ""))
 
+                args.append(str(python_max))
+
                 # TODO: this positional `args` handling is getting out of hand, leverage _write_sensitive_values?
 
                 try:
                     log.info(f"Submitting job with command {args!r}")
-                    script_output = subprocess.check_output(args, stderr=subprocess.STDOUT, universal_newlines=True)
+                    d = dict(**os.environ)
+                    d["YARN_CONTAINER_RUNTIME_DOCKER_IMAGE"] = image_name
+                    script_output = subprocess.check_output(args, stderr=subprocess.STDOUT, universal_newlines=True, env=d)
                     log.info(f"Submitted job, output was: {script_output}")
                 except CalledProcessError as e:
                     log.error(f"Submitting job failed, output was: {e.stdout}", exc_info=True)