+++ openeogeotrellis/deploy/batch_job_metadata.py
@@ -360,11 +360,12 @@ def _get_tracker(tracker_id=""):
 
 def _get_tracker_metadata(tracker_id="") -> dict:
     tracker = _get_tracker(tracker_id)
+    usage = {}
+    all_links = []
 
     if tracker is not None:
         tracker_results = tracker.asDict()
 
-        usage = {}
         pu = tracker_results.get("Sentinelhub_Processing_Units", None)
         if pu is not None:
             usage["sentinelhub"] = {"value": pu, "unit": "sentinelhub_processing_unit"}
@@ -389,4 +390,12 @@ def _get_tracker_metadata(tracker_id="") -> dict:
                 for link in all_links
             ]
 
-        return dict_no_none(usage=usage if usage != {} else None, links=all_links)
+
+    from openeogeotrellis.metrics_tracking import global_tracker
+    python_metrics = global_tracker().as_dict()
+    sar_backscatter_errors = python_metrics.pop("orfeo_backscatter_soft_errors",0)
+    sar_backscatter_total = python_metrics.pop("orfeo_backscatter_execution_counter",0)
+    usage =  { **usage, **{name: {"value": value, "unit": "count"} for name, value in python_metrics.items()}}
+    if sar_backscatter_total > 0:
+        usage["sar_backscatter_soft_errors"] = {"value": sar_backscatter_errors / sar_backscatter_total, "unit": "fraction"}
+    return dict_no_none(usage=usage if usage != {} else None, links=all_links)