+++ tests/conftest.py
@@ -4,6 +4,7 @@ import os
 import shutil
 import sys
 import typing
+import urllib
 from datetime import datetime
 from pathlib import Path
 from typing import Union
@@ -359,6 +360,46 @@ def vault() -> Vault:
 def urllib_mock() -> UrllibMocker:
     with UrllibMocker().patch() as mocker:
         yield mocker
+
+
+class UrllibPoolManagerMocker(UrllibMocker):
+    @contextlib.contextmanager
+    def patch(self):
+        with mock.patch('urllib3.poolmanager.PoolManager.request', new=self._request):
+            yield self
+
+    def _request(self, method, url, fields=None, headers={}, **urlopen_kw):
+        for match_url in [url, self._drop_query_string(url)]:
+            key = (method, match_url)
+            if key in self.response_callbacks:
+                return self.response_callbacks[key](urllib.request.Request(url, method=method, headers=headers))
+        return self.Response(code=404, msg="Not Found")
+
+
+@pytest.fixture
+def urllib_poolmanager_mock() -> UrllibPoolManagerMocker:
+    with UrllibPoolManagerMocker().patch() as mocker:
+        yield mocker
+
+
+class UrllibAndRequestMocker:
+    def __init__(self, urllib_mock, requests_mock, urllib_poolmanager_mock):
+        self.urllib_mock = urllib_mock
+        self.requests_mock = requests_mock
+        self.urllib_poolmanager_mock = urllib_poolmanager_mock
+
+    def get(self, href, data):
+        code = 200
+        self.urllib_mock.get(href, data, code)
+        self.urllib_poolmanager_mock.get(href, data, code)
+        if isinstance(data, str):
+            data = data.encode("utf-8")
+        self.requests_mock.get(href, content=data)
+
+
+@pytest.fixture
+def urllib_and_request_mock(urllib_mock, requests_mock, urllib_poolmanager_mock) -> UrllibAndRequestMocker:
+    yield UrllibAndRequestMocker(urllib_mock, requests_mock, urllib_poolmanager_mock)
 
 
 TEST_AWS_REGION_NAME = "eu-central-1"