+++ tests/test_load_stac.py
@@ -8,6 +8,7 @@ from openeo_driver.errors import OpenEOApiException
 from openeo_driver.utils import EvalEnv
 
 from openeogeotrellis.load_stac import extract_own_job_info, load_stac
+from tests.data import get_test_data_file
 
 
 @pytest.mark.parametrize("url, user_id, job_info_id",
@@ -35,40 +36,9 @@ def test_extract_own_job_info(url, user_id, job_info_id):
         assert job_info.id == job_info_id
 
 
-def test_property_filter_from_parameter(urllib_mock, requests_mock):
-    stac_root_url = "https://stac.test"
-    stac_collection_url = f"{stac_root_url}/collections/collection"
-
-    urllib_mock.get(stac_collection_url, data=json.dumps({
-        "type": "Collection",
-        "stac_version": "1.0.0",
-        "id": "collection",
-        "description": "collection",
-        "license": "unknown",
-        "extent": {
-            "spatial": {"bbox": [[-180, -90, 180, 90]]},
-            "temporal": {"interval": [[None, None]]},
-        },
-        "links": [{
-            "rel": "root",
-            "href": stac_root_url,
-        }],
-    }))
-
-    catalog_response = {
-        "type": "Catalog",
-        "stac_version": "1.0.0",
-        "id": "stac.test",
-        "description": "stac.test",
-        "links": [],
-        "conformsTo": [
-            "https://api.stacspec.org/v1.0.0-rc.1/item-search",
-            "https://api.stacspec.org/v1.0.0-rc.3/item-search#filter"
-        ],
-    }
-
-    urllib_mock.get(stac_root_url, data=json.dumps(catalog_response))
-    requests_mock.get(stac_root_url, json=catalog_response)
+def test_property_filter_from_parameter(urllib_and_request_mock, requests_mock):
+    stac_api_root_url = "https://stac.test"
+    stac_collection_url = f"{stac_api_root_url}/collections/collection"
 
     def feature_collection(request, _) -> dict:
         assert request.qs["filter-lang"] == ["cql2-text"]
@@ -81,7 +51,7 @@ def test_property_filter_from_parameter(urllib_mock, requests_mock):
             "features": [],
         }
 
-    search_mock = requests_mock.get(f"{stac_root_url}/search", json=feature_collection)
+    search_mock = _mock_stac_api(urllib_and_request_mock, requests_mock, stac_api_root_url, stac_collection_url, feature_collection)
 
     properties = {
         "product_tile": {
@@ -112,3 +82,142 @@ def test_property_filter_from_parameter(urllib_mock, requests_mock):
         )
 
     assert search_mock.called
+
+
+def test_dimensions(urllib_and_request_mock, requests_mock):
+    stac_api_root_url = "https://stac.test"
+    stac_collection_url = f"{stac_api_root_url}/collections/collection"
+
+    stac_item = json.loads(
+        get_test_data_file("stac/issue609-api-temporal-bound-exclusive/item01.json")
+        .read_text()
+        .replace("asset01.tiff", f"file://{get_test_data_file('binary/load_stac/collection01/asset01.tif').absolute()}")
+    )
+
+    _mock_stac_api(
+        urllib_and_request_mock,
+        requests_mock,
+        stac_api_root_url,
+        stac_collection_url,
+        feature_collection={
+            "type": "FeatureCollection",
+            "features": [stac_item],
+        },
+    )
+
+    data_cube = load_stac(
+        url=stac_collection_url,
+        load_params=LoadParameters(),
+        env=EvalEnv({"pyramid_levels": "highest"}),
+        layer_properties={},
+        batch_jobs=None,
+    )
+
+    assert {"x", "y", "t", "bands"} <= set(data_cube.metadata.dimension_names())
+
+
+@pytest.fixture
+def jvm_mock():
+    with mock.patch("openeogeotrellis.load_stac.get_jvm") as get_jvm:
+        jvm_mock = get_jvm.return_value
+
+        raster_layer = mock.MagicMock()
+        jvm_mock.geopyspark.geotrellis.TemporalTiledRasterLayer.return_value = raster_layer
+        raster_layer.layerMetadata.return_value = """{
+            "crs": "EPSG:4326",
+            "cellType": "uint8",
+            "bounds": {"minKey": {"col":0, "row":0}, "maxKey": {"col": 1, "row": 1}},
+            "extent": {"xmin": 0,"ymin": 0, "xmax": 1,"ymax": 1},
+            "layoutDefinition": {
+                "extent": {"xmin": 0, "ymin": 0,"xmax": 1,"ymax": 1},
+                "tileLayout": {"layoutCols": 1, "layoutRows": 1, "tileCols": 256, "tileRows": 256}
+            }
+        }"""
+
+        yield jvm_mock
+
+
+@pytest.mark.parametrize(
+    ["band_names", "resolution"],
+    [
+        (["AOT_10m"], 10.0),
+        (["WVP_20m"], 20.0),
+        (["WVP_60m"], 60.0),
+        (["AOT_10m", "WVP_20m"], 10.0)
+    ],
+)
+def test_data_cube_resolution_matches_requested_bands(
+    urllib_and_request_mock, requests_mock, jvm_mock, band_names, resolution
+):
+    stac_api_root_url = "https://stac.test"
+    stac_collection_url = f"{stac_api_root_url}/collections/collection"
+
+    features = json.loads(get_test_data_file("stac/issue1043-api-proj-code/FeatureCollection.json").read_text())
+
+    _mock_stac_api(
+        urllib_and_request_mock,
+        requests_mock,
+        stac_api_root_url,
+        stac_collection_url,
+        feature_collection=features,
+    )
+
+    factory_mock = jvm_mock.org.openeo.geotrellis.file.PyramidFactory
+    cellsize_mock = jvm_mock.geotrellis.raster.CellSize
+
+    data_cube = load_stac(
+        stac_collection_url,
+        load_params=LoadParameters(bands=band_names),
+        env=EvalEnv({"pyramid_levels": "highest"}),
+        layer_properties={},
+        batch_jobs=None,
+        apply_lcfm_improvements=True,
+    )
+
+    # TODO: how to check the actual argument to PyramidFactory()?
+    factory_mock.assert_called_once()
+    cellsize_mock.assert_called_once_with(resolution, resolution)
+    assert data_cube.metadata.spatial_extent["crs"] == "EPSG:32636"
+
+
+def _mock_stac_api(urllib_and_request_mock, requests_mock, stac_api_root_url, stac_collection_url, feature_collection):
+    urllib_and_request_mock.get(
+        stac_collection_url,
+        data=json.dumps(
+            {
+                "type": "Collection",
+                "stac_version": "1.0.0",
+                "id": "collection",
+                "description": "collection",
+                "license": "unknown",
+                "extent": {
+                    "spatial": {"bbox": [[-180, -90, 180, 90]]},
+                    "temporal": {"interval": [[None, None]]},
+                },
+                "links": [
+                    {
+                        "rel": "root",
+                        "href": stac_api_root_url,
+                    }
+                ],
+            }
+        ),
+    )
+
+    catalog_response = {
+        "type": "Catalog",
+        "stac_version": "1.0.0",
+        "id": "stac.test",
+        "description": "stac.test",
+        "links": [],
+        "conformsTo": [
+            "https://api.stacspec.org/v1.0.0-rc.1/item-search",
+            "https://api.stacspec.org/v1.0.0-rc.3/item-search#filter",
+        ],
+    }
+
+    urllib_and_request_mock.get(stac_api_root_url, data=json.dumps(catalog_response))
+    requests_mock.get(stac_api_root_url, json=catalog_response)
+
+    search_mock = requests_mock.get(f"{stac_api_root_url}/search", json=feature_collection)
+    return search_mock