+++ tests/test_views.py
@@ -514,7 +514,7 @@ class TestBatchJobs:
             assert batch_job_args[5] == job_metadata.name
             assert batch_job_args[8] == TEST_USER
             assert batch_job_args[9] == api.api_version
-            assert batch_job_args[10:16] == ['8G', '2G', '3G', '5', '2', '2G']
+            assert batch_job_args[10:16] == ['8G', '2G', '3072m', '5', '2', '2G']
             assert batch_job_args[16:21] == [
                 'default', 'false', '[]',
                 "__pyfiles__/custom_processes.py,foolib.whl", '100'
@@ -962,7 +962,7 @@ class TestBatchJobs:
             assert batch_job_args[5] == job_metadata.name
             assert batch_job_args[8] == TEST_USER
             assert batch_job_args[9] == api.api_version
-            assert batch_job_args[10:16] == ['3g', '11g', '3G', '5', '4', '10G']
+            assert batch_job_args[10:16] == ['3g', '11g', '3072m', '5', '4', '10G']
             assert batch_job_args[16:21] == [
                 'somequeue', 'false', '[]',
                 '__pyfiles__/custom_processes.py,foolib.whl', '100'